restaurantapis/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cloudinary==1.44.1
Django==6.0
django-ckeditor==6.7.3
django-cors-headers==4.3.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-oauth-toolkit==1.7.1
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==12.1.0
PyMySQL==1.1.2
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
